requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.5.1
pymongo==4.8.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
passlib==1.7.4

# ── Pydantic / settings ───────────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic[email]

# ── Quantum ML ────────────────────────────────────────────────────────────────
pennylane==0.37.0

# ── Classical ML ─────────────────────────────────────────────────────────────
scikit-learn==1.5.1
xgboost==2.1.1
shap==0.46.0

# ── OCR ───────────────────────────────────────────────────────────────────────
pytesseract==0.3.13
paddleocr==2.8.1
Pillow==10.4.0

# ── PDF ───────────────────────────────────────────────────────────────────────
reportlab==4.2.2
weasyprint==62.3
PyPDF2==3.0.1
pdfplumber==0.11.4

# ── LLM ───────────────────────────────────────────────────────────────────────
google-generativeai==0.7.2
openai==1.40.0

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
cryptography==43.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
hypothesis==6.112.1
httpx==0.27.0
